check update/requirements.txt
PySide6>=6.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
APScheduler>=3.10.0
python-dateutil>=2.8.2
SQLAlchemy>=2.0.0
pillow>=10.0.0
PyQt6>=6.4.0
pycryptodome>=3.15.0 